requirements.txt
imageio
numpy
torch
Pillow
transformers